requirements.txt
gradio>=4.25
pandas
requests
python-dotenv
Pillow>=9.0.0 